src/requirements.txt
flask==1.1.2
Flask-SocketIO==4.3.1
Flask-PyMongo
requests
jsonschema
pytest